requirements.txt
Django==3.0.5
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
pathlib==1.0.1
psycopg2-binary==2.8.5
pytest-django==3.9.0
python-decouple==3.4
six==1.14.0